chore(allow-scripts): break out config.js from setup.js
