+++ packages/allow-scripts/src/index.js
@@ -1,19 +1,13 @@
 // @ts-check
 
 const {
-  areBinsBlocked,
-  editPackageJson,
   getOptionsForBin,
   loadAllPackageConfigurations,
   setDefaultConfiguration,
-  writeRcFile,
-} = require('./setup.js')
-const {
-  printPackagesList,
-  printMissingPoliciesIfAny
-} = require('./report.js')
+} = require('./config.js')
+const { areBinsBlocked, editPackageJson, writeRcFile } = require('./setup.js')
+const { printPackagesList, printMissingPoliciesIfAny } = require('./report.js')
 const runAllowedPackages = require('./runAllowedPackages.js')
-
 
 module.exports = {
   getOptionsForBin,
